+++ src/renderer.js
@@ -1,0 +1,31 @@
+// This file is required by the index.html file
+// and will be executed in the renderer process
+// for that window.
+
+// Renderer process for Naval Tactical Simulator
+const { ipcRenderer } = require('electron');
+
+console.log('Renderer process started');
+
+// Handle simulation events
+ipcRenderer.on('new-simulation', () => {
+  console.log('Creating new simulation');
+  // TODO: Initialize new simulation
+});
+
+ipcRenderer.on('start-simulation', () => {
+  console.log('Starting simulation');
+  // TODO: Start simulation
+});
+
+ipcRenderer.on('pause-simulation', () => {
+  console.log('Pausing simulation');
+  // TODO: Pause simulation
+});
+
+ipcRenderer.on('reset-simulation', () => {
+  console.log('Resetting simulation');
+  // TODO: Reset simulation
+});
+
+// You can add more renderer process code here